00_TrackerApp/backend/requirements.txt
autopep8==1.5.6
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.1
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.4.9
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
uvicorn==0.13.4
wheel==0.36.2
